feat: add `.promise` to the helpers
